backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.1
redis==5.0.1
celery==5.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
aiohttp==3.9.1
pandas==2.1.4
numpy==1.25.2
python-dotenv==1.0.0

# Broker integrations
requests==2.31.0
websockets==12.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-xdist==3.5.0
pytest-mock==3.12.0
pytest-benchmark==4.0.0
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1
bandit==1.7.5
coverage==7.3.2